RuboCop対応のため、UserDecoratorをモジュールに切り分けた
